Split sequential scaling in a different file
