+++ contrib/pdl/src/__init__.py
@@ -1,0 +1,11 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""
+The module holds the generic PDL I/O processor and a sequential scaling processor.
+"""
+
+# First Party
+from contrib.pdl.src.io.pdl_io import PdlInputOutputProcessor as PdlInputOutputProcessor
+from contrib.pdl.src.io.sequential_scaling import (  # noqa: F401
+    SequentialScalingInputOutputProcessor as SequentialScalingInputOutputProcessor,
+)
